// src/schema/settings.schema.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Configuration settings for users and organizations on Sourcegraph.
 */
export interface Settings {
  /**
   * Saved search queries
   */
  "search.savedQueries"?: {
    /**
     * Unique key for this query in this file
     */
    key: string;
    /**
     * Description of this saved query
     */
    description: string;
    /**
     * Query string
     */
    query: string;
    /**
     * Show this saved query on the homepage
     */
    showOnHomepage?: boolean;
    /**
     * Notify the owner of this configuration file when new results are available
     */
    notify?: boolean;
    /**
     * Notify Slack via the organization's Slack webhook URL when new results are available
     */
    notifySlack?: boolean;
  }[];
  /**
   * Predefined search scopes
   */
  "search.scopes"?: SearchScope[];
  /**
   * Named groups of repositories that can be referenced in a search query using the repogroup: operator.
   */
  "search.repositoryGroups"?: {
    [k: string]: string[];
  };
  "notifications.slack"?: SlackNotificationsConfig;
  /**
   * An array (often with just one element) of messages to display at the top of all pages, including for unauthenticated users. Users may dismiss a message (and any message with the same string value will remain dismissed for the user).
   *
   * Markdown formatting is supported.
   *
   * Usually this setting is used in global and organization settings. If set in user settings, the message will only be displayed to that user. (This is useful for testing the correctness of the message's Markdown formatting.)
   *
   * MOTD stands for "message of the day" (which is the conventional Unix name for this type of message).
   */
  motd?: string[];
  /**
   * The Sourcegraph extensions to use. Enable an extension by adding a property `"my/extension": true` (where `my/extension` is the extension ID). Override a previously enabled extension and disable it by setting its value to `false`.
   */
  extensions?: {
    /**
     * `true` to enable the extension, `false` to disable the extension (if it was previously enabled)
     */
    [k: string]: boolean;
  };
  [k: string]: any;
}
export interface SearchScope {
  /**
   * A unique identifier for the search scope.
   *
   * If set, a scoped search page is available at https://[sourcegraph-hostname]/search/scope/ID, where ID is this value.
   */
  id?: string;
  /**
   * The human-readable name for this search scope
   */
  name: string;
  /**
   * The query string of this search scope
   */
  value: string;
  /**
   * A description for this search scope
   */
  description?: string;
}
/**
 * Configuration for sending notifications to Slack.
 */
export interface SlackNotificationsConfig {
  /**
   * The Slack webhook URL used to post notification messages to a Slack channel. To obtain this URL, go to: https://YOUR-WORKSPACE-NAME.slack.com/apps/new/A0F7XDUAZ-incoming-webhooks
   */
  webhookURL: string;
}