// src/schema/site.schema.d.ts
export interface SiteConfiguration {
// ...
  licenseKey?: string;
// ...
  "auth.providers"?: (BuiltinAuthProvider | SAMLAuthProvider | OpenIDConnectAuthProvider | HTTPHeaderAuthProvider)[];
// ...
}
// ...
export interface GitHubConnection {
  /**
   * URL of a GitHub instance, such as https://github.com or https://github-enterprise.example.com.
   */
  url: string;
  /**
   * The type of Git URLs to use for cloning and fetching Git repositories on this GitHub instance.
   *
   * If "http", Sourcegraph will access GitLab repositories using Git URLs of the form http(s)://github.com/myteam/myproject.git (using https: if the GitHub instance uses HTTPS).
   *
   * If "ssh", Sourcegraph will access GitHub repositories using Git URLs of the form git@github.com:myteam/myproject.git. See the documentation for how to provide SSH private keys and known_hosts: https://about.sourcegraph.com/docs/config/repositories#repositories-that-need-https-or-ssh-authentication.
   */
  gitURLType?: "http" | "ssh";
  /**
   * A GitHub personal access token with repo and org scope.
   */
  token: string;
  /**
   * TLS certificate of a GitHub Enterprise instance.
   */
  certificate?: string;
  /**
   * An array of repository "owner/name" strings specifying which GitHub or GitHub Enterprise repositories to mirror on Sourcegraph.
   */
  repos?: string[];
  /**
   * An array of strings specifying which GitHub or GitHub Enterprise repositories to mirror on Sourcegraph. The valid values are:
   *
   * - `public` mirrors all public repositories for GitHub Enterprise and is the equivalent of `none` for GitHub
   *
   * - `affiliated` mirrors all repositories affiliated with the configured token's user:
   * 	- Private repositories with read access
   * 	- Public repositories owned by the user or their orgs
   * 	- Public repositories with write access
   *
   * - `none` mirrors no repositories (except those specified in the `repos` configuration property or added manually)
   *
   * If multiple values are provided, their results are unioned.
   *
   * If you need to narrow the set of mirrored repositories further (and don't want to enumerate the set in the "repos" configuration property), create a new bot/machine user on GitHub or GitHub Enterprise that is only affiliated with the desired repositories.
   */
  repositoryQuery?: ("public" | "affiliated" | "none")[];
  /**
   * The pattern used to generate the corresponding Sourcegraph repository path for a GitHub or GitHub Enterprise repository. In the pattern, the variable "{host}" is replaced with the GitHub host (such as github.example.com), and "{nameWithOwner}" is replaced with the GitHub repository's "owner/path" (such as "myorg/myrepo").
   *
   * For example, if your GitHub Enterprise URL is https://github.example.com and your Sourcegraph URL is https://src.example.com, then a repositoryPathPattern of "{host}/{nameWithOwner}" would mean that a GitHub repository at https://github.example.com/myorg/myrepo is available on Sourcegraph at https://src.example.com/github.example.com/myorg/myrepo.
   */
  repositoryPathPattern?: string;
  /**
   * Defines whether repositories from this GitHub instance should be enabled and cloned when they are first seen by Sourcegraph. If false, the site admin must explicitly enable GitHub repositories (in the site admin area) to clone them and make them searchable on Sourcegraph. If true, they will be enabled and cloned immediately (subject to rate limiting by GitHub); site admins can still disable them explicitly, and they'll remain disabled.
   */
  initialRepositoryEnablement?: boolean;
}
export interface GitLabConnection {
  /**
   * URL of a GitLab instance, such as https://gitlab.example.com or (for GitLab.com) https://gitlab.com.
   */
  url: string;
  /**
   * A GitLab personal access token with "api" scope.
   */
  token: string;
  /**
   * The type of Git URLs to use for cloning and fetching Git repositories on this GitLab instance.
   *
   * If "http", Sourcegraph will access GitLab repositories using Git URLs of the form http(s)://gitlab.example.com/myteam/myproject.git (using https: if the GitLab instance uses HTTPS).
   *
   * If "ssh", Sourcegraph will access GitLab repositories using Git URLs of the form git@example.gitlab.com:myteam/myproject.git. See the documentation for how to provide SSH private keys and known_hosts: https://about.sourcegraph.com/docs/config/repositories#repositories-that-need-https-or-ssh-authentication.
   */
  gitURLType?: "http" | "ssh";
  /**
   * TLS certificate of a GitLab instance.
   */
  certificate?: string;
  /**
   * An array of strings specifying which GitLab projects to mirror on Sourcegraph. Each string is a URL query string for the GitLab projects API, such as "?membership=true&search=foo".
   *
   * The query string is passed directly to GitLab to retrieve the list of projects. The special string "none" can be used as the only element to disable this feature. Projects matched by multiple query strings are only imported once. See https://docs.gitlab.com/ee/api/projects.html#list-all-projects for available query string options.
   */
  projectQuery?: string[];
  /**
   * The pattern used to generate a the corresponding Sourcegraph repository path for a GitLab project. In the pattern, the variable "{host}" is replaced with the GitLab URL's host (such as gitlab.example.com), and "{pathWithNamespace}" is replaced with the GitLab project's "namespace/path" (such as "myteam/myproject").
   *
   * For example, if your GitLab is https://gitlab.example.com and your Sourcegraph is https://src.example.com, then a repositoryPathPattern of "{host}/{pathWithNamespace}" would mean that a GitLab project at https://gitlab.example.com/myteam/myproject is available on Sourcegraph at https://src.example.com/gitlab.example.com/myteam/myproject.
   */
  repositoryPathPattern?: string;
  /**
   * Defines whether repositories from this GitLab instance should be enabled and cloned when they are first seen by Sourcegraph. If false, the site admin must explicitly enable GitLab repositories (in the site admin area) to clone them and make them searchable on Sourcegraph. If true, they will be enabled and cloned immediately (subject to rate limiting by GitLab); site admins can still disable them explicitly, and they'll remain disabled.
   */
  initialRepositoryEnablement?: boolean;
}
// ...
export interface BitbucketServerConnection {
  /**
   * URL of a Bitbucket Server instance, such as https://bitbucket.example.com.
   */
  url: string;
  /**
   * A Bitbucket Server personal access token with Read scope. Create one at https://[your-bitbucket-hostname]/plugins/servlet/access-tokens/add.
   *
   * For Bitbucket Server instances that don't support personal access tokens (Bitbucket Server version 5.4 and older), specify user-password credentials in the "username" and "password" fields.
   */
  token?: string;
  /**
   * The username to use when authenticating to the Bitbucket Server instance. Also set the corresponding "password" field.
   *
   * For Bitbucket Server instances that support personal access tokens (Bitbucket Server version 5.5 and newer), it is recommended to provide a token instead (in the "token" field).
   */
  username?: string;
  /**
   * The password to use when authenticating to the Bitbucket Server instance. Also set the corresponding "username" field.
   *
   * For Bitbucket Server instances that support personal access tokens (Bitbucket Server version 5.5 and newer), it is recommended to provide a token instead (in the "token" field).
   */
  password?: string;
  /**
   * The type of Git URLs to use for cloning and fetching Git repositories on this Bitbucket Server instance.
   *
   * If "http", Sourcegraph will access Bitbucket Server repositories using Git URLs of the form http(s)://bitbucket.example.com/scm/myproject/myrepo.git (using https: if the Bitbucket Server instance uses HTTPS).
   *
   * If "ssh", Sourcegraph will access Bitbucket Server repositories using Git URLs of the form ssh://git@example.bitbucket.com/myproject/myrepo.git. See the documentation for how to provide SSH private keys and known_hosts: https://about.sourcegraph.com/docs/config/repositories#repositories-that-need-https-or-ssh-authentication.
   */
  gitURLType?: "http" | "ssh";
  /**
   * TLS certificate of a Bitbucket Server instance.
   */
  certificate?: string;
  /**
   * The pattern used to generate the corresponding Sourcegraph repository path for a Bitbucket Server repository.
   *
   *  - "{host}" is replaced with the Bitbucket Server URL's host (such as bitbucket.example.com)
   *  - "{projectKey}" is replaced with the Bitbucket repository's parent project key (such as "PRJ")
   *  - "{repositorySlug}" is replaced with the Bitbucket repository's slug key (such as "my-repo").
   *
   * For example, if your Bitbucket Server is https://bitbucket.example.com and your Sourcegraph is https://src.example.com, then a repositoryPathPattern of "{host}/{projectKey}/{repositorySlug}" would mean that a Bitbucket Server repository at https://bitbucket.example.com/projects/PRJ/repos/my-repo is available on Sourcegraph at https://src.example.com/bitbucket.example.com/PRJ/my-repo.
   */
  repositoryPathPattern?: string;
  /**
   * Whether or not personal repositories should be excluded or not. When true, Sourcegraph will ignore personal repositories it may have access to. See https://about.sourcegraph.com/docs/config/repositories/#excluding-personal-repositories for more information. Default: false.
   */
  excludePersonalRepositories?: boolean;
  /**
   * Defines whether repositories from this Bitbucket Server instance should be enabled and cloned when they are first seen by Sourcegraph. If false, the site admin must explicitly enable Bitbucket Server repositories (in the site admin area) to clone them and make them searchable on Sourcegraph. If true, they will be enabled and cloned immediately (subject to rate limiting by Bitbucket Server); site admins can still disable them explicitly, and they'll remain disabled.
   */
  initialRepositoryEnablement?: boolean;
}
// ...
export interface Repository {
  /**
   * Type of the version control system for this repository, such as "git"
   */
  type?: "git";
  /**
   * Clone URL for the repository, such as git@example.com:my/repo.git
   */
  url: string;
  /**
   * Display path on Sourcegraph for the repository, such as my/repo
   */
  path: string;
  links?: {
    /**
     * URL specifying where to view the repository at an external location e.g. "https://example.com/myrepo"
     */
    repository?: string;
    /**
     * URL template for specifying how to link to commits at an external location. Use "{commit}" as a placeholder for a given commit ID e.g. "https://example.com/myrepo/view-commit/{commit}"
     */
    commit?: string;
    /**
     * URL template for specifying how to link to paths at an external location. Use "{path}" as a placeholder for a given path and "{rev}" as a placeholder for a given revision e.g. "https://example.com/myrepo@{rev}/browse/{path}"
     */
    tree?: string;
    /**
     * URL template for specifying how to link to files at an external location. Use "{path}" as a placeholder for a given path and "{rev}" as a placeholder for a given revision e.g. "https://example.com/myrepo@{rev}/browse/{path}"
     */
    blob?: string;
  };
}
/**
 * Configures the builtin username-password authentication provider.
 */
export interface BuiltinAuthProvider {
  type: string;
  /**
   * Allows new visitors to sign up for accounts. The sign-up page will be enabled and accessible to all visitors.
   *
   * SECURITY: If the site has no users (i.e., during initial setup), it will always allow the first user to sign up and become site admin **without any approval** (first user to sign up becomes the admin).
   */
  allowSignup?: boolean;
}